+++ disassembler/src/style/s.rs
@@ -1,7 +1,7 @@
 pub fn s_common(name: &str, raw_inst: u64) -> String {
     let rs2 = (raw_inst >> 8) & 0b11111;
     let rs1 = (raw_inst >> 13) & 0b111;
-    let imm = (raw_inst >> 16) & 0xffffffff;
+    let imm = ((raw_inst >> 16) & 0xffffffff) as i32;
 
     format!("{} r{}[{}] =  r{}", name, rs1, imm, rs2)
 }
